nrf: add support for nrf52805, nrf52811, nrf52820
