+++ embassy-nrf/Cargo.toml
@@ -11,11 +11,13 @@ defmt-info = [ ]
 defmt-warn = [ ]
 defmt-error = [ ]
 
-52810 = ["nrf52810-pac"]
-52811 = ["nrf52811-pac"]
-52832 = ["nrf52832-pac"]
-52833 = ["nrf52833-pac"]
-52840 = ["nrf52840-pac"]
+nrf52805 = ["nrf52805-pac"]
+nrf52810 = ["nrf52810-pac"]
+nrf52811 = ["nrf52811-pac"]
+nrf52820 = ["nrf52820-pac"]
+nrf52832 = ["nrf52832-pac"]
+nrf52833 = ["nrf52833-pac"]
+nrf52840 = ["nrf52840-pac"]
 
 
 [dependencies]
@@ -30,10 +32,12 @@ cortex-m = "0.7.1"
 embedded-hal    = { version = "0.2.4" }
 embedded-dma    = { version = "0.1.2" }
 futures     = { version = "0.3.5", default-features = false }
+critical-section = "0.2.1"
 
+nrf52805-pac  = { version = "0.1.0", optional = true, features = [ "rt" ], git = "https://github.com/Dirbaio/nrf52805-pac"}
 nrf52810-pac  = { version = "0.9.0", optional = true, features = [ "rt" ]}
 nrf52811-pac  = { version = "0.9.1", optional = true, features = [ "rt" ]}
+nrf52820-pac  = { version = "0.1.0", optional = true, features = [ "rt" ], git = "https://github.com/Dirbaio/nrf52820-pac"}
 nrf52832-pac  = { version = "0.9.0", optional = true, features = [ "rt" ]}
 nrf52833-pac  = { version = "0.9.0", optional = true, features = [ "rt" ]}
 nrf52840-pac  = { version = "0.9.0", optional = true, features = [ "rt" ]}
-critical-section = "0.2.1"
